git-svn-id: http://svn.code.sf.net/p/distributedcl/code/trunk@313 cb344d40-e590-48b6-843f-7a98a4fc1ff3

+++ src/libdistributedcl/message/msg_program.cpp
@@ -89,7 +89,7 @@ void dcl_message< msgCreateProgramWithBinary >::create_request( void* payload_pt
 
     for( uint16_t i = 0; i < devices_.size(); ++i )
     {
-        *lengths = host_to_network( lengths_[ i ] );
+        *lengths = host_to_network( static_cast<uint32_t>( lengths_[ i ] ) );
         ++lengths;
     }
